+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Card
   * Fix an issue where a tokenize card errors are masked with a JSON parsing error
+* PayPal
+    * Add `shouldOfferCredit`property to `PayPalCheckoutRequest`. Enables merchants to offer PayPal Credit for Checkout requests.
   
 ## 5.18.0 (2025-10-01)